+++ src/commands/borealis-pg/tunnel.test.ts
@@ -415,7 +415,7 @@ describe('secure tunnel command', () => {
 
       errorListener({code: 'EADDRINUSE'})
 
-      expect(ctx.stderr).to.contain(`Local port ${customPgPort} is already in use`)
+      expect(ctx.stderr).to.contain(`Local port ${customPgPort} is not available`)
       verify(mockNodeProcessType.exit(1)).once()
     })